some shit

+++ src/engine/enginebufferscalerubberband.h
@@ -20,7 +20,7 @@ class EngineBufferScaleRubberBand : public EngineBufferScale {
                                     double* pTempoRatio,
                                     double* pPitchRatio);
 
-    virtual void setSampleRate(int iSampleRate);
+    virtual void setSampleRate(double iSampleRate);
 
     // Read and scale buf_size samples from the provided RAMAN.
     CSAMPLE* getScaled(unsigned long buf_size);
@@ -29,7 +29,7 @@ class EngineBufferScaleRubberBand : public EngineBufferScale {
     void clear();
 
     // Reset RubberBand library with new samplerate.
-    void initializeRubberBand(int iSampleRate);
+    void initializeRubberBand(double iSampleRate);
   private:
     void deinterleaveAndProcess(const CSAMPLE* pBuffer, size_t frames, bool flush);
     size_t retrieveAndDeinterleave(CSAMPLE* pBuffer, size_t frames);
